Autopilot dragonfly: add ArduCopter interface
With the ArduCopter interface code can now be
generated. However, there will several double
data type variables in the generated code.
If all variables in the autopilot parameters
struct are converted to singles, there are
still double data type variables in the generated
code.
In rtwtypes.h all "double" can be replaced with
"single" as a temporary workaround.
It was not tested yet, if ArduCopter compiles
when integrating this code.
In QuadcopterSimModel QuadcopterSimModel_Dragonfly
the controller inputs "waypoints" and
"trigger_wp_update" where interchanged after
LADAC update.