Resolve initialization error when using FlightGear visualization
At the first start of the simulation, the flying machines have
been flung upwards, if FlightGear was used as visualization.
The problem was known:
The flight dynamics simulation in Simulink receives the current
ground altitude from FlightGear - which was necessary to make
landings look realistic in FlightGear. However, the initial
ground altitude sent by FlightGear was the last previously
determined ground altitude. Updating the ground altitude
sometimes takes 1 second or longer.
The problem was solved by setting the initial FlightGear
position to the default position used in Simulink.
However, the problem should still occur if the ground
elevation in the previous simulation was higher than the
initial position.