Bug fixes: Flight modes altitude hold and loiter
The altitude hold flight mode examples did not
run because the connections to the altitude
flight mode were wrong.
Moreover, in both altitude flight mode and
loiter flight mode a sensor filtering was
performed outside of the flight mode block.
These filters had to be removed because the
filtering is now performed inside of the flight
mode block.